fix(config): add support for 'gpt-3.5-turbo-16k' model in configValidators
The 'configValidators' object in the 'config.ts' file was updated to include support for the 'gpt-3.5-turbo-16k' model. Previously, only 'gpt-3.5-turbo' and 'gpt-4' models were supported. This change allows users to use the 'gpt-3.5-turbo-16k' model as well. The validation message was also updated to reflect this change.